Changed more ints to bigints for v3 (#923)
* Add v3 migration - filevault_escrow

* Change some timestamps to bigints

* Changed more ints to bigints

+++ app/modules/smart_stats/migrations/2017_12_19_195801_smart_stats.php
@@ -29,7 +29,7 @@ class SmartStats extends Migration
             $table->string('smart_is')->nullable();
             $table->integer('error_count')->nullable();
             $table->integer('error_poh')->nullable();
-            $table->integer('timestamp')->nullable();
+            $table->bigInteger('timestamp')->nullable();
             $table->bigInteger('raw_read_error_rate')->nullable();
             $table->bigInteger('throughput_performance')->nullable();
             $table->bigInteger('spin_up_time')->nullable();
